plugins/JmComicPlugin/requirements.txt
# 插件依赖项
jmcomic
Pillow
